reduced code redundancy caused by the additional feature #89 not finished work
